+++ src/backend/joanie/tests/signature/backends/lex_persona/test_create_workflow.py
@@ -7,6 +7,8 @@ import responses
 from joanie.signature import exceptions
 from joanie.signature.backends import get_signature_backend
 
+# pylint: disable=protected-access
+
 
 @override_settings(
     JOANIE_SIGNATURE_BACKEND="joanie.signature.backends.lex_persona.LexPersonaBackend",
@@ -136,7 +138,7 @@ class LexPersonaBackendCreateWorkflowTestCase(TestCase):
 
         self.assertEqual(
             str(context.exception),
-            "Cannot create a signature procedure.",
+            "Lex Persona: Cannot create a signature procedure.",
         )
         self.assertEqual(len(responses.calls), 1)
         self.assertEqual(responses.calls[0].request.url, api_url)